feat: removed drawer widget
