--- app/cdap/components/common/DrawerWidget/IconStore/Underline.tsx
@@ -1,34 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import React from 'react';
-
-// This image is an SVG and is not available in MUI Icons Store
-export const Underline = () => {
-  return (
-    <svg
-      width="67"
-      height="3"
-      viewBox="0 0 67 3"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      data-testid="underline-icon"
-    >
-      <path d="M0 0.530273H50L53 2.5318H3L0 0.530273Z" fill="#2196F3" />
-      <path d="M54 0.530273H63.5L66.5 2.5318H57L54 0.530273Z" fill="#2196F3" />
-    </svg>
-  );
-};
